Introduce patroni.postgresql.mpp package
